Farmart Update (Ignore)

+++ platform/core/base/src/Supports/Database.php
@@ -3,11 +3,16 @@
 namespace Botble\Base\Supports;
 
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\File;
 
 class Database
 {
     public static function restoreFromPath(string $pathToSqlFile, string $connection = null): void
     {
+        if (! File::exists($pathToSqlFile) || File::size($pathToSqlFile) < 1024) {
+            return;
+        }
+
         DB::purge($connection);
         DB::connection()->setDatabaseName(DB::getDatabaseName());
         DB::getSchemaBuilder()->dropAllTables();
